streamlit/requirements.txt
streamlit>=1.35
streamlit-drawable-canvas>=0.9.3
Pillow>=10.0
streamlit-keyup>=0.2.0
